Move chc_local to chc crate
