+++ crates/holochain/src/conductor/chc.rs
@@ -6,10 +6,7 @@ use once_cell::sync::Lazy;
 use std::{collections::HashMap, sync::Arc};
 use url::Url;
 
-mod chc_local;
-pub use chc_local::*;
-
-pub use holochain_chc::*;
+pub use holochain_chc::{chc_http::*, chc_local::*, *};
 
 /// Storage for the local CHC implementations
 pub static CHC_LOCAL_MAP: Lazy<parking_lot::Mutex<HashMap<CellId, Arc<ChcLocal>>>> =
@@ -50,4 +47,275 @@ fn chc_local(keystore: MetaLairClient, cell_id: CellId) -> ChcImpl {
 
 fn chc_remote(url: Url, keystore: MetaLairClient, cell_id: &CellId) -> ChcImpl {
     Arc::new(ChcHttp::new(url, keystore, cell_id))
+}
+
+#[cfg(test)]
+mod tests {
+
+    use hdk::prelude::*;
+    use holochain_chc::*;
+    use holochain_conductor_api::conductor::ConductorConfig;
+    use holochain_types::record::SignedActionHashedExt;
+    use holochain_wasm_test_utils::TestWasm;
+
+    use crate::{
+        conductor::{
+            api::error::ConductorApiError,
+            chc::{CHC_LOCAL_MAGIC_URL, CHC_LOCAL_MAP},
+            error::ConductorError,
+        },
+        sweettest::*,
+    };
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn simple_chc_sync() {
+        use holochain::test_utils::inline_zomes::simple_crud_zome;
+
+        let config = ConductorConfig {
+            chc_url: Some(url2::Url2::parse(CHC_LOCAL_MAGIC_URL)),
+            ..Default::default()
+        };
+        let mut conductor = SweetConductor::from_config(config).await;
+
+        let (dna_file, _, _) = SweetDnaFile::unique_from_inline_zomes(simple_crud_zome()).await;
+
+        let (cell,) = conductor
+            .setup_app("app", &[dna_file])
+            .await
+            .unwrap()
+            .into_tuple();
+
+        let cell_id = cell.cell_id();
+        let agent = cell_id.agent_pubkey().clone();
+
+        let top_hash = {
+            let mut dump = conductor.dump_full_cell_state(cell_id, None).await.unwrap();
+            assert_eq!(dump.source_chain_dump.records.len(), 3);
+            dump.source_chain_dump.records.pop().unwrap().action_address
+        };
+
+        let izc = InitZomesComplete {
+            author: agent.clone(),
+            timestamp: Timestamp::now(),
+            action_seq: 3,
+            prev_action: top_hash,
+        };
+        let new_action = ActionHashed::from_content_sync(Action::InitZomesComplete(izc));
+        let new_action = SignedActionHashed::sign(&conductor.keystore(), new_action)
+            .await
+            .unwrap();
+        let new_action_hash = new_action.action_address().clone();
+        let new_record = Record::new(new_action, None);
+
+        {
+            // add some data to the local CHC
+            let m = CHC_LOCAL_MAP.lock();
+            let chc = m.get(cell_id).unwrap();
+            let records = chc.clone().get_record_data(None).await.unwrap();
+            assert_eq!(records.len(), 3);
+            chc.clone().add_records(vec![new_record]).await.unwrap();
+        }
+
+        // Check that a sync picks up the new action
+        conductor
+            .raw_handle()
+            .chc_sync(cell_id.clone(), None)
+            .await
+            .unwrap();
+
+        let dump = conductor.dump_full_cell_state(cell_id, None).await.unwrap();
+        assert_eq!(dump.source_chain_dump.records.len(), 4);
+        assert_eq!(
+            dump.source_chain_dump
+                .records
+                .last()
+                .unwrap()
+                .action_address,
+            new_action_hash,
+        );
+    }
+
+    // TODO: run this remotely too
+    #[tokio::test(flavor = "multi_thread")]
+    async fn multi_conductor_chc_sync() {
+        holochain_trace::test_run();
+
+        let mut config = SweetConductorConfig::standard().no_dpki();
+        // config.chc_url = Some(url2::Url2::parse("http://127.0.0.1:40845/"));
+        config.chc_url = Some(url2::Url2::parse(CHC_LOCAL_MAGIC_URL));
+        let mut conductors = SweetConductorBatch::from_config(4, config).await;
+
+        let (dna_file, _, _) = SweetDnaFile::unique_from_test_wasms(vec![TestWasm::Create]).await;
+
+        // All conductors share the same known agent, already installed in the test_keystore
+        let agent = SweetAgents::alice();
+
+        let (c0,) = conductors[0]
+            .setup_app_for_agent("app", agent.clone(), &[dna_file.clone()])
+            .await
+            .unwrap()
+            .into_tuple();
+
+        let cell_id = c0.cell_id();
+
+        // Install two apps with ignore_genesis_failure and one without
+        let mk_payload = |ignore: bool| {
+            let agent = agent.clone();
+            let dna_file = dna_file.clone();
+            async move {
+                let mut payload =
+                    get_install_app_payload_from_dnas("app", Some(agent), &[(dna_file, None)])
+                        .await;
+                payload.ignore_genesis_failure = ignore;
+                payload
+            }
+        };
+
+        let install_result_1 = conductors[1]
+            .raw_handle()
+            .install_app_bundle(mk_payload(true).await)
+            .await;
+        let install_result_2 = conductors[2]
+            .raw_handle()
+            .install_app_bundle(mk_payload(true).await)
+            .await;
+        let install_result_3 = conductors[3]
+            .raw_handle()
+            .install_app_bundle(mk_payload(false).await)
+            .await;
+
+        // It's not ideal to match on a string, but it seems like the only option:
+        // - The pattern involves Boxes which are impossible to match on
+        // - The error types are not PartialEq, so cannot be constructed and tested for equality
+
+        dbg!(&install_result_1);
+        dbg!(&install_result_2);
+        dbg!(&install_result_3);
+
+        regex::Regex::new(
+            r#".*ChcHeadMoved\("genesis", InvalidChain\((\d+), ActionHash\([a-zA-Z0-9-_]+\)\)\).*"#,
+        )
+        .unwrap()
+        .captures(&format!("{:?}", install_result_1))
+        .unwrap();
+        // TODO: check sequence and hash
+
+        assert_eq!(
+            format!("{:?}", install_result_1),
+            format!("{:?}", install_result_2)
+        );
+        assert_eq!(
+            format!("{:?}", install_result_2),
+            format!("{:?}", install_result_3)
+        );
+
+        assert!(conductors[1]
+            .get_app_info(&"app".into())
+            .await
+            .unwrap()
+            .is_some());
+        assert!(conductors[2]
+            .get_app_info(&"app".into())
+            .await
+            .unwrap()
+            .is_some());
+
+        // This one will not have app info, since it was installed without `ignore_genesis_failure`
+        assert_eq!(
+            conductors[3].get_app_info(&"app".into()).await.unwrap(),
+            None
+        );
+
+        // TODO: sync conductors 1 and 2 to match conductor 0
+        conductors[1]
+            .raw_handle()
+            .chc_sync(cell_id.clone(), None)
+            .await
+            .unwrap();
+        conductors[2]
+            .raw_handle()
+            .chc_sync(cell_id.clone(), None)
+            .await
+            .unwrap();
+
+        // Sync is not possible since the installation was rolled back and the cell was removed
+        assert!(matches!(
+            conductors[3]
+                .raw_handle()
+                .chc_sync(cell_id.clone(), None)
+                .await,
+            Err(ConductorApiError::ConductorError(ConductorError::CellMissing(id))) if id == *cell_id
+        ));
+
+        let dump1 = conductors[1]
+            .dump_full_cell_state(cell_id, None)
+            .await
+            .unwrap();
+
+        assert_eq!(dump1.source_chain_dump.records.len(), 3);
+
+        let c1: SweetCell = conductors[1].get_sweet_cell(cell_id.clone()).unwrap();
+        let c2: SweetCell = conductors[2].get_sweet_cell(cell_id.clone()).unwrap();
+
+        let _: ActionHash = conductors[0]
+            .call(&c0.zome(TestWasm::Create), "create_entry", ())
+            .await;
+
+        conductors[1].enable_app("app".into()).await.unwrap();
+        conductors[2].enable_app("app".into()).await.unwrap();
+
+        // This should fail and require triggering a CHC sync
+        let hash1: Result<ActionHash, _> = conductors[1]
+            .call_fallible(&c1.zome(TestWasm::Create), "create_entry", ())
+            .await;
+
+        dbg!(&hash1);
+
+        regex::Regex::new(
+            r#".*ChcHeadMoved\("SourceChain::flush", InvalidChain\((\d+), ActionHash\([a-zA-Z0-9-_]+\).*"#
+        ).unwrap().captures(&format!("{:?}", hash1)).unwrap();
+        // TODO: check sequence and hash
+
+        // This should trigger a CHC sync
+        let hash2: Result<ActionHash, _> = conductors[2]
+            .call_fallible(&c2.zome(TestWasm::Create), "create_entry", ())
+            .await;
+
+        assert_eq!(format!("{:?}", hash1), format!("{:?}", hash2));
+
+        conductors[1]
+            .raw_handle()
+            .chc_sync(cell_id.clone(), None)
+            .await
+            .unwrap();
+
+        conductors[2]
+            .raw_handle()
+            .chc_sync(cell_id.clone(), None)
+            .await
+            .unwrap();
+
+        let dump0 = conductors[0]
+            .dump_full_cell_state(cell_id, None)
+            .await
+            .unwrap();
+        let dump1 = conductors[1]
+            .dump_full_cell_state(cell_id, None)
+            .await
+            .unwrap();
+        let dump2 = conductors[2]
+            .dump_full_cell_state(cell_id, None)
+            .await
+            .unwrap();
+
+        assert_eq!(dump0.source_chain_dump.records.len(), 6);
+        assert_eq!(
+            dump0.source_chain_dump.records,
+            dump1.source_chain_dump.records
+        );
+        assert_eq!(
+            dump1.source_chain_dump.records,
+            dump2.source_chain_dump.records
+        );
+    }
 }
