+++ crates/holochain/src/core/sys_validate/error.rs
@@ -1,4 +1,3 @@
-use derive_more::Display;
 use holochain_conductor_services::DpkiServiceError;
 use std::convert::TryFrom;
 
@@ -186,65 +185,4 @@ impl ValidationOutcome {
         }
         matches!(self, Self::DepMissingFromDht(_) | Self::DpkiAgentMissing(_))
     }
-}
-
-/// Context information for an invalid action to make it easier to trace in errors.
-#[derive(Error, Debug, Display, PartialEq, Eq)]
-#[display(
-    fmt = "{} - with context seq={}, action_hash={:?}, action=[{}]",
-    source,
-    seq,
-    action_hash,
-    action_display
-)]
-pub struct PrevActionError {
-    #[source]
-    pub source: PrevActionErrorKind,
-    pub seq: u32,
-    pub action_hash: ActionHash,
-    pub action_display: String,
-}
-
-impl<A: ChainItem> From<(PrevActionErrorKind, &A)> for PrevActionError {
-    fn from((inner, action): (PrevActionErrorKind, &A)) -> Self {
-        PrevActionError {
-            source: inner,
-            seq: action.seq(),
-            action_hash: action.get_hash().clone().into(),
-            action_display: action.to_display(),
-        }
-    }
-}
-
-impl From<(PrevActionErrorKind, Action)> for PrevActionError {
-    fn from((inner, action): (PrevActionErrorKind, Action)) -> Self {
-        PrevActionError {
-            source: inner,
-            seq: action.action_seq(),
-            action_hash: action.to_hash(),
-            action_display: format!("{}", action),
-        }
-    }
-}
-
-#[derive(Error, Debug, PartialEq, Eq)]
-pub enum PrevActionErrorKind {
-    #[error("The previous action hash specified in an action doesn't match the actual previous action. Seq: {0}")]
-    HashMismatch(u32),
-    #[error("Root of source chain must be Dna")]
-    InvalidRoot,
-    #[error("Root of source chain must have a timestamp greater than the Dna's origin_time")]
-    InvalidRootOriginTime,
-    #[error("No more actions are allowed after a chain close")]
-    ActionAfterChainClose,
-    #[error("Previous action sequence number {1} != ({0} - 1)")]
-    InvalidSeq(u32, u32),
-    #[error("Action is not the first, so needs previous action")]
-    MissingPrev,
-    #[error("The previous action's timestamp is not before the current action's timestamp: {0:?} >= {1:?}")]
-    Timestamp(Timestamp, Timestamp),
-    #[error("The previous action's author does not match the current action's author: {0} != {1}")]
-    Author(AgentPubKey, AgentPubKey),
-    #[error("It is invalid for these two actions to be paired with each other. context: {0}, actions: {1:?}")]
-    InvalidSuccessor(String, Box<(Action, Action)>),
 }
