+++ crates/holochain_chc/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "holochain_chc"
 version = "0.1.0-dev.1"
-description = "TODO"
+description = "Defines the Chain Head Coordinator (CHC) API for Holochain and provides an HTTP client implementation."
 license = "Apache-2.0"
 homepage = "https://github.com/holochain/holochain"
 documentation = "https://docs.rs/holochain_chc"
@@ -20,6 +20,7 @@ holochain_serialized_bytes = "=0.0.55"
 holochain_types = { path = "../holochain_types", version = "^0.4.0-dev.20" }
 one_err = "0.0.8"
 must_future = "0.1.1"
+parking_lot = "0.12"
 reqwest = { version = "0.12", features = ["json", "native-tls-vendored"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_bytes = "0.11.12"
